@@ -60,47 +71,49 @@
     "karma-mocha-reporter": "^2.0.4",
     "karma-phantomjs-launcher": "^1.0.1",
     "karma-sauce-launcher": "^1.1.0",
-    "karma-source-map-support": "^1.1.0",
+    "karma-source-map-support": "^1.2.0",
     "karma-sourcemap-loader": "^0.3.6",
-    "karma-webpack": "^2.0.1",
+    "karma-webpack": "^2.0.3",
     "mocha": "^3.0.1",
     "npm-run-all": "^4.0.0",
     "phantomjs-prebuilt": "^2.1.7",
     "rimraf": "^2.5.3",
-    "rollup": "^0.40.0",
-    "rollup-plugin-babel": "^1.0.0",
+    "rollup": "^0.48.2",
+    "rollup-plugin-babel": "^3.0.2",
     "rollup-plugin-memory": "^2.0.0",
-    "rollup-plugin-node-resolve": "^2.0.0",
-    "sinon": "^1.17.4",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "sinon": "^2.2.0",
     "sinon-chai": "^2.8.0",
     "typescript": "^2.2.2",
     "uglify-js": "^2.7.5",
-    "webpack": "^1.13.1"
+    "webpack": "^2.4.1"
   },
   "eslintConfig": {
     "extends": "./config/eslint-config.js"
   },
   "files": [
     "devtools",
+    "debug",
     "src",
     "dist",
     "devtools.js",
     "devtools.js.map",
+    "debug.js",
+    "debug.js.map",
     "typings.json"
   ],
   "greenkeeper": {
     "ignore": [
-      "rollup-plugin-babel",
-      "babel",
+      "babel-cli",
       "babel-core",
       "babel-eslint",
       "babel-loader",
-      "babel-runtime",
-      "jscodeshift"
+      "jscodeshift",
+      "rollup-plugin-babel"
     ]
   },
   "homepage": "https://github.com/developit/preact",
-  "jsnext:main": "src/preact.js",
+  "jsnext:main": "dist/preact.esm.js",
   "keywords": [
     "preact",
     "react",
