@@ -16,7 +16,7 @@ export class AuthServiceProviderService {
           const data = { email: username, password: password };
           return this.api.post('login', { email: username, password: password })
               .map(user => {
-                  console.log(user);
+                  // console.log(user);
                   // login successful if there's a jwt token in the response
                   if (user && user.token) {
                       // store user details and jwt token in local storage to keep user logged in between page refreshes
