@@ -5,32 +5,38 @@ import {Subject} from 'rxjs/Subject';
 import {ApiProvider} from '../libraries/api';
 import {AlertProviderService} from './alert.service';
 import {isPlatformBrowser} from '@angular/common';
+import {AuthServiceProviderService} from './auth.service';
 
 @Injectable()
 export class CartService {
 
-    constructor(private api: ApiProvider, private alertService: AlertProviderService, @Inject(PLATFORM_ID) private platformId: Object) {
+    constructor(private api: ApiProvider, private alertService: AlertProviderService, @Inject(PLATFORM_ID) private platformId: Object,
+                private authService: AuthServiceProviderService) {
         // console.log('constructor', localStorage.token);
-        if(isPlatformBrowser(this.platformId)) {
-            if (localStorage.getItem('token')) {
-                this.init();
-            }
-        }
+        this.checkAuth();
     }
 
     private cartSubject = new Subject<any>();
     cart_data: any;
     CartState = this.cartSubject.asObservable();
 
+    checkAuth() {
+        if (isPlatformBrowser(this.platformId)) {
+            if (localStorage.getItem('token')) {
+                this.init();
+            }
+        }
+    }
+
     addProduct(_product: any) {
         // console.log(_product);
-        return this.api.post('insertc',
+        return this.api.post('cart',
             { product_id: _product, qty: 1});
         // console.log(this.CartState);
     }
 
     removeProduct(id: any) {
-        return this.api.delete(`delete_item_cart/${id}`)
+        return this.api.delete(`cart/${id}`)
             .subscribe( res => {
                 this.init();
             });
@@ -38,11 +44,11 @@ export class CartService {
 
     getAllcart_data(): Observable <any> {
         // console.log('from service', this.httpOptions);
-        return this.api.get('viewc');
+        return this.api.get('cart');
     }
 
     init() {
-        this.api.get('viewc')
+        this.api.get('cart')
             .subscribe(res => {
                 this.cart_data = res;
                 this.cartSubject.next(<any>{loaded: true, cart_data: this.cart_data});
@@ -50,15 +56,15 @@ export class CartService {
     }
 
     deleteCart() {
-        return this.api.delete('deletec');
+        return this.api.delete('cart');
     }
 
     addToCart(id: any, qty: any, price: any) {
-        return this.api.post('insertc', { product_id: id, qty: qty, price: price });
+        return this.api.post('cart', { product_id: id, qty: qty, price: price });
     }
 
     getAddressData() {
-        return this.api.get('viewuadd');
+        return this.api.get('address');
     }
 
     validateCart() {
@@ -67,7 +73,7 @@ export class CartService {
 
     updateCart(product_id: any, qty: any) {
 
-        return this.api.put('update_cart_qty',  { product_id: product_id, qty: qty});
+        return this.api.put('cart',  { product_id: product_id, qty: qty});
     }
 
 }
